Added v2 of the code.
This html and css presents a totally different project compared to the original committed ones. A bigger delve into how the effect Glassmorphism works and how to better use it. The point of this code is to study how to implement Glassmorphism into a project. I didn't spend any time trying to make it responsive or anything like that.